Fix warnings caused by signed and unsigned int comparisson

+++ tests/libc/pc_hw/co80/scputc.c
@@ -8,7 +8,7 @@ main(void)
   int i,j;
   ScreenClear();
   for (i=0; i<20; i++)
-    for (j=0; j<sizeof(str); j++)
+    for (j=0; j<(int)sizeof(str); j++)
       ScreenPutChar(str[j], i, i+j, i);
   return 0;
 }
